backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.34.0,<0.35.0

# Database
sqlalchemy>=2.0.0,<2.1.0
psycopg2-binary>=2.9.0,<3.0.0
alembic>=1.14.0,<2.0.0

# Authentication (Clerk JWT verification)
pyjwt>=2.10.0,<3.0.0
cryptography>=44.0.0,<45.0.0

# Validation
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0

# CORS
python-multipart>=0.0.20,<0.1.0

# PDF Generation
reportlab>=4.2.0,<5.0.0

# Email
resend>=2.0.0,<3.0.0

# Payments
stripe>=12.0.0,<13.0.0

# Environment
python-dotenv>=1.0.0,<2.0.0

# HTTP Client
httpx>=0.27.0,<0.29.0

# Testing
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.25.0,<0.26.0
